Move hooks to separate files.
